more fixes

+++ source/world/tile/entity/DispenserTileEntity.cpp
@@ -10,7 +10,7 @@ ItemStack DispenserTileEntity::removeRandomItem()
 	int var1 = -1;
 	int var2 = 1;
 
-	for (int var3 = 0; var3 < m_items.size(); ++var3) {
+	for (int var3 = 0; var3 < (int)m_items.size(); ++var3) {
 		if (!m_items[var3].isEmpty() && m_random.nextInt(var2) == 0) {
 			var1 = var3;
 			++var2;
